+++ src/common/game.d.ts
@@ -68,7 +68,7 @@ export interface GameScore {
   cells: number,
   mines: number,
   level?: number,
-  playMode?: PlayMode,
+  playMode?: PlayMode, // backwards compatibility
   mode?: PlayMode,
   effort: {
     least: number,