+++ src/common/update.ts
@@ -1,6 +1,8 @@
 import { ScoreItem } from './game.d'
+import { DEFAULTS } from './defaults'
 import { SCORE_RADIX } from './constants'
 import { capitalise } from './functions'
+import { AppConfig, PlayMode } from './app.d'
 
 /**
  Add updates at the end
@@ -32,27 +34,42 @@ const updater = () => {
 
   const victoriesStorage = localStorage.getItem('mv-victories')
   if (victoriesStorage) {
+    const playModeNames = Object.values(PlayMode)
+
+    /* CONFIGURATION */
+
+    const configStorage = localStorage.getItem('mv-config')
+    if (configStorage) {
+      const config: AppConfig = { ...JSON.parse(configStorage), ...DEFAULTS }
+
+      const updPlayMode = capitalise(config.PLAY_MODE)
+      config.PLAY_MODE = PlayMode[updPlayMode as keyof typeof PlayMode] || PlayMode.NORMAL
+
+      localStorage.setItem('mv-config', JSON.stringify(config))
+      console.log('Configuration updated.')
+    }
+
+    /* SCORE-LIST */
+
     const scores: ScoreItem[] = JSON.parse(victoriesStorage)
     const converted = scores.map((s: ScoreItem) => {
       // one-bit values in code property
       const playModeString = capitalise(s.game?.playMode) || '--'
-      const playModeNames = ['Normal', 'Bare', 'Sharp']
       const playModeNumber = playModeNames.findIndex(el => el === playModeString)
-      const newCode = s.code.substring(0,3) + Math.max(playModeNumber, 0) + s.code.substring(3)
-      // format score items
+      const boundModeNumber = Math.max(playModeNumber, 0)
+      const newCode = s.code.substring(0,3) + boundModeNumber + s.code.substring(3)
       s.code = newCode
-      if (Object.hasOwn(s.game, 'playMode')) {
-        if (playModeNumber > 0) {
-          s.game.mode = playModeString
-        }
-        delete s.game?.playMode
+      // format score items
+      if (playModeNumber > 0) {
+        s.game.mode = playModeString
       }
+      delete s.game?.playMode
       s.game.level = parseInt(s.code.charAt(2), SCORE_RADIX)
       return s
     })
 
     localStorage.setItem('mv-won-games', JSON.stringify(converted))
-//    localStorage.removeItem('mv-victories')
+    localStorage.removeItem('mv-victories')
     console.log('Scorelist updated to replay with right playmode.')
   }
 }