+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 - Logic for updating data in storage
 - Console tools overview in markdown
 - Toggle descriptions for navigation and option buttons, show by default
-- Write level always and mode when selected in game object of scoreItem
+- Write level always and mode when selected in game section of scoreItem
 
 ### Changed
 - Levels refined with intermediate values, level values have been doubled and localStorage is adapted accordingly.